refactor: progress banner and move some utils to lib
